Decode script
